feat: se comparan archivos con un archivo externo y se corrigen los posibles errores
