+++ src/chat/package.json
@@ -4,7 +4,7 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "start:intento-2": "nodemon --exec babel-node res/indexchat.js",
+    "start:intento-2": "nodemon --exec babel-node src/indexchat.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "presets": [
