+++ src/chat/src/public/index.html
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
 
-  <link rel="stylesheet" href="./scss/stylechat.scss">
+  <link rel="stylesheet" href="./scss/stylechat.css">
   <title>chat</title>
 </head>
 
@@ -119,18 +119,12 @@
   <footer>
     
   </footer>
-
-  <!-- este script de jquery nos permite trabajar con un codigo de java un poco mas acortado -->
   <script
   src="https://code.jquery.com/jquery-3.7.1.min.js"
   integrity="sha256-/JqT3SQfawRcv/BIHPThkBvs0OEvtFFmqPF/lYI/Cxo="
   crossorigin="anonymous"></script>
-
-  <!-- funcionalidad de boobstrap -->
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
-  <!-- nos enlaza con el servidor que tenemos definido de socket.io -->
   <script type="module" src="socket.io/socket.io.js"></script>
-  <!-- funcionalidad desde el controlador principal de la interface -->
   <script type="module" src="./js/mainchat.js"></script>
   
 </body>
