+++ src/chat/src/public/js/mainchat.js
@@ -1,10 +1,10 @@
-const socket=io()
 
+const socket=io()
 $(function(){
+
     //este es el encargado de enviar los datos
     //esta es la conexion de sockets de el cliente
     const socket=io()
-
 
     //obteniendo los elementos del doom desde la interface 
     const $messageform =$(`#message-form`);
@@ -13,6 +13,7 @@ $(function(){
     console.log($messagebox)
     const $chat=$(`#chat`)
     console.log($chat)
+
 
     //eventos
     $messageform.submit(e=>{
@@ -29,8 +30,7 @@ $(function(){
 
     })
 
-
-
-
-
 })
+
+
+
