+++ src/chat/src/public/scss/stylechat.css
@@ -251,3 +251,6 @@ body{
       width: 100%; /* Ocupa todo el ancho disponible */
     }
   }
+
+
+
